requirements.txt
chainlit==2.6.0
fastmcp==2.8.0
langchain==0.3.26
langgraph==0.5.1
mcp==1.9.2
pydantic==2.11.7
python-dotenv==1.1.0
typing_extensions==4.14.1
langchain-openai==0.3.27
uv
elasticsearch-mcp-server
